# desktop-app/openapi/stocktrader_desktop.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StockTrader | desktop",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/backtest/optimize": {
      "post": {
        "tags": [
          "StockTrader"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/optimize-jobs": {
      "post": {
        "tags": [
          "StockTrader"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOptimizeJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/optimize-jobs/{id}": {
      "get": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/optimize-jobs/{id}/settings": {
      "post": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOptimizeJobSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/optimize-jobs/{id}/apply-result": {
      "post": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyOptimizeJobResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/optimize-jobs/{id}/cancel": {
      "post": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/optimize-jobs/{id}/pause": {
      "post": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/optimize-jobs/{id}/resume": {
      "post": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/optimize-jobs/{id}/results": {
      "get": {
        "tags": [
          "StockTrader"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "top",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyOptimizeJobResultRequest": {
        "type": "object",
        "properties": {
          "resultId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "CreateOptimizeJobRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "chunkSize": {
            "type": "integer",
            "format": "int32"
          },
          "maxDurationHours": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "maxTestedCombinations": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "topResultsToKeep": {
            "type": "integer",
            "format": "int32"
          },
          "rankBy": {
            "type": "string"
          },
          "continuousMode": {
            "type": "boolean"
          },
          "autoApplyBestResult": {
            "type": "boolean"
          },
          "autoApplyMinTrades": {
            "type": "integer",
            "format": "int32"
          },
          "optimizeRequest": {
            "$ref": "#/components/schemas/OptimizeRequest"
          }
        }
      },
...
      "DataSource": {
        "enum": [
          "Alpaca",
          "Polygon",
          "Yahoo",
          "LsSecurities",
          null
        ]
      },
...
      "OptimizeParams": {
        "type": "object",
        "properties": {
          "atrStopMultiplier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "atrTargetMultiplier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "maxHoldingBars": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "trailingAtr": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "partialProfitR": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "ruleParamOverrides": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/RuleParamRange"
            }
          },
          "entryLogicOptions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "requireBullRegimeOptions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "boolean"
            }
          },
          "entryModeOptions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "sizingModeOptions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "exitLogicOptions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "timeFrameOptions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "defaultAllocationPercent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "circuitBreakerConsecutiveLossLimit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "circuitBreakerCooldownBars": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "circuitBreakerMaxDrawdownPercent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "reentryCooldownAfterLoss": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "reentryCooldownAfterWin": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "portfolioMaxPositions": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "portfolioMaxSinglePercent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "portfolioMaxEntriesPerDay": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParamRange"
              }
            ]
          },
          "ruleFieldOverrides": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/RuleFieldRange"
            }
          }
        }
      },
      "OptimizeRequest": {
        "type": "object",
        "properties": {
          "basePattern": {
            "$ref": "#/components/schemas/StrategyDocument"
          },
          "symbols": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          },
          "initialCapital": {
            "type": "number",
            "format": "double"
          },
          "dataSource": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DataSource"
              }
            ]
          },
          "timeFrame": {
            "$ref": "#/components/schemas/TimeFrame"
          },
          "optimizeParams": {
            "$ref": "#/components/schemas/OptimizeParams"
          },
          "rankBy": {
            "type": "string"
          },
          "maxResults": {
            "type": "integer",
            "format": "int32"
          },
          "maxCombinations": {
            "type": "integer",
            "format": "int32"
          },
          "oosPercent": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ParamRange": {
        "type": "object",
        "properties": {
          "min": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "max": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "step": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "values": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        }
      },
...
      "RuleParamRange": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string"
          },
          "ruleIndex": {
            "type": "integer",
            "format": "int32"
          },
          "paramKey": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            }
          }
        }
      },
...
      "StrategyDocument": {
        "type": "object",
        "properties": {
          "storedStrategyId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "documentVersion": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "entryRulesJson": {
            "type": "string"
          },
          "entryLogic": {
            "type": "string"
          },
          "requireBullRegime": {
            "type": "boolean"
          },
          "atrStopMultiplier": {
            "type": "number",
            "format": "double"
          },
          "atrTargetMultiplier": {
            "type": "number",
            "format": "double"
          },
          "maxHoldingBars": {
            "type": "integer",
            "format": "int32"
          },
          "trailingAtr": {
            "type": "number",
            "format": "double"
          },
          "partialProfitR": {
            "type": "number",
            "format": "double"
          },
          "useWeightTiers": {
            "type": "boolean"
          },
          "weightTiersJson": {
            "type": "string"
          },
          "defaultAllocationPercent": {
            "type": "number",
            "format": "double"
          },
          "exitRulesJson": {
            "type": "string"
          },
          "exitRulesLogic": {
            "type": "string"
          },
          "exitGroupsJson": {
            "type": "string"
          },
          "exitGroupsLogic": {
            "type": "string"
          },
          "scalingRulesJson": {
            "type": "string"
          },
          "timeFilterJson": {
            "type": "string"
          },
          "circuitBreakerJson": {
            "type": "string"
          },
          "reentryJson": {
            "type": "string"
          },
          "portfolioRulesJson": {
            "type": "string"
          },
          "entryGroupsJson": {
            "type": "string"
          },
          "entryGroupsLogic": {
            "type": "string"
          },
          "dynamicExitJson": {
            "type": "string"
          },
          "entryMode": {
            "type": "string"
          },
          "timeFrame": {
            "$ref": "#/components/schemas/TimeFrame"
          },
          "sizingMode": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "enableLiveTrading": {
            "type": "boolean"
          }
        }
      },
      "TimeFrame": {
        "enum": [
          "OneMinute",
          "FiveMinute",
          "FifteenMinute",
          "Daily",
          "Weekly"
        ]
      },
      "UpdateOptimizeJobSettingsRequest": {
        "type": "object",
        "properties": {
          "autoApplyBestResult": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "autoApplyMinTrades": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}